# Faster processing of array comparisons with constant offsets