--- a/src/Analyser/MutatingScope.php
+++ b/src/Analyser/MutatingScope.php
@@ -4428,6 +4428,10 @@ final class MutatingScope implements Scope
 
 		if ($originalExprType->equals($nativeType)) {
 			$newType = TypeCombinator::intersect($type, $originalExprType);
+			if ($newType->isConstantScalarValue()->yes() && $newType->equals($nativeType)) {
+				// don't add the same type over and over again to improve performance
+				return $this;
+			}
 			return $this->specifyExpressionType($expr, $newType, $newType, TrinaryLogic::createYes());
 		}