step debugging thru the example made me realize that most of the time we were adding the same constant types to the scope again. This PR detects this situation and does a fast exit to prevent unnecessary scope creations.

<img width="1430" alt="grafik" src="https://github.com/user-attachments/assets/03567e0f-0808-4bb8-b681-6f438d090090" />


---

before this PR
```
time php bin/phpstan analyze bug-12800.php --debug
  13.40s user 0.18s system 99% cpu 13.637 total
```

after this PR
```
time php bin/phpstan analyze bug-12800.php --debug
  2.84s user 0.13s system 99% cpu 2.974 total
```

<img width="570" alt="grafik" src="https://github.com/user-attachments/assets/2ac00c49-9735-4970-9990-3331f677d467" />


closes https://github.com/phpstan/phpstan/issues/12800